swap profile and storypoints to utils
